Adding functionality to add a game for a user and remove a game

+++ db_service/flaskr/mail.py
@@ -6,11 +6,12 @@ from flaskr.db import get_db
 
 bp = Blueprint('mail', __name__, url_prefix='/mail')
 
-@bp.route('/add', methods=['GET', 'POST'])
+@bp.route('/add', methods=['GET', 'POST', 'OPTIONS'])
 def add_user_game():
     if request.method == 'POST':
-        user = request.form["user"]
-        game = request.form["game"]
+        data = request.get_json()
+        user = data["user"]
+        game = data["game"]
         db = get_db()
         error = None
 
@@ -50,6 +51,10 @@ def add_user_game():
                 return response
         
         return (error, 500)
+    
+    if request.method == 'OPTIONS':
+        response = "success"
+        return response
 
 @bp.route('/get', methods=['GET', 'POST'])
 def get_user_game():
@@ -86,42 +91,48 @@ def get_user_game():
 
 @bp.route('/delete', methods=['DELETE'])
 def delete_user_game():
-    user = request.form["user"]
-    game = request.form["game"]
-    db = get_db()
-    error = None
+    if request.method == 'DELETE':
+        data = request.get_json()
+        user = data["user"]
+        game = data["game"]
+        db = get_db()
+        error = None
 
-    if not user:
+        if not user:
             error = 'username is required'
-    elif not game:
-        error = 'game is required'
+        elif not game:
+            error = 'game is required'
 
-    if error is None:
-        game_id = db.execute("SELECT id from game where name = ?", (game,)
-        ).fetchone()
+        if error is None:
+            game_id = db.execute("SELECT id from game where name = ?", (game,)
+            ).fetchone()
+            
+            if game_id is None:
+                error = "game not found in database"
+            
+            user_id = db.execute("SELECT id from user where username = ?", (user,)
+            ).fetchone()
+
+            if user_id is None:
+                error = "user not found in database"
+
+        if error is None:
+            game_id = game_id["id"]
+            user_id = user_id["id"]
+
+            try:
+                db.execute(
+                    "DELETE FROM users_games WHERE user_id = ? and game_id = ?",
+                    (user_id, game_id),
+                )
+                db.commit()
+            except db.IntegrityError:
+                error = f"Failed to delete relationship {user, game} from database"
+            else:
+                return jsonify( { "success": True } ), 200
         
-        if game_id is None:
-            error = "game not found in database"
-        
-        user_id = db.execute("SELECT id from user where username = ?", (user,)
-        ).fetchone()
-
-        if user_id is None:
-            error = "user not found in database"
-
-    if error is None:
-        game_id = game_id["id"]
-        user_id = user_id["id"]
-
-        try:
-            db.execute(
-                "DELETE FROM users_games WHERE user_id = ? and game_id = ?",
-                (user_id, game_id),
-            )
-            db.commit()
-        except db.IntegrityError:
-            error = f"Failed to delete relationship {user, game} from database"
-        else:
-            return jsonify( { "success": True } ), 200
+        return (error, 500)
     
-    return (error, 500)
+    if request.method == 'OPTIONS':
+        response = "success"
+        return response
